vcl: Adds backend tests to check for correct rendering
Collection of test that check rendering of VCL backend. This works
by rendering primitives to the VirtualDevice, outputing the result
to a Bitmap and checking the rendering by comparing pixels. This
should warn when the primitive is drawn to a wrong position or if
it is the wrong size or color. It also test some rendering tricks
like XOR drawing.

Currently this is not implemented as a CPPUnit test but separately,
as its own executable file so we can check the results visually.
The idea is also to add the tests into LO for the user to test or
automatically to test the backend on first start (especially
useful to test the OpenGL backend as some drivers could have
unexpected behaviour).

Change-Id: I2a565b9986ca68850e2752466e3cd5fc71991ba6