Simplify user name option GUI: Just a single "name" field
See https://www.kalzumeus.com/2010/06/17/falsehoods-programmers-believe-about-names/

And it makes the code simpler. Win-win.

It is just a single full name that is stored in ODF and OOXML anyway.

Take care of migrating separate given names, father's names and
surnames from older user profiles if modifying the data. Handle the
"Eastern" (as the code calls it) case, too.

Change-Id: Iebf59f224995f722669f844d03053a55c088bc6e