ucb: webdav-curl: tdf#101094 (32): Use a partial GET as a fall-back for HEAD...
... partial GET means a GET with no data returned, to test for
resource 'real' availability in case HEAD fails or is disabled.
At  the same time disable use of HEAD method on the resource
for the whole access procedure.

This change is needed to manage web servers that don't permit the
use of either HEAD or OPTIONS methods.
To accomodate pure web sites that enable only GET to fetch contents.

[ port of commit f9b978c27f41a746461f54097402313fb4b282df ]

Change-Id: Ie0e2b6edd0860cbf704082500b91631543e7da38
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/123488
Tested-by: Michael Stahl <michael.stahl@allotropia.de>
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>