tdf#101094 (32): Use a partial GET as a fall-back for HEAD...
... partial GET means a GET with no data returned, to test for
resource 'real' availability in case HEAD fails or is disabled.
At  the same time disable use of HEAD method on the resource
for the whole access procedure.

This change is needed to manage web servers that don't permit the
use of either HEAD or OPTIONS methods.
To accomodate pure web sites that enable only GET to fetch contents.

Change-Id: If52316ddd6ce637fb7d2ef518423de246ee2dcfb
Reviewed-on: https://gerrit.libreoffice.org/28992
Tested-by: Giuseppe Castagno <giuseppe.castagno@acca-esse.eu>
Reviewed-by: Giuseppe Castagno <giuseppe.castagno@acca-esse.eu>