tdf#96248 delete SfxPoolItems with Which Id >= 4000
Which Ids > 4999 (SFX_WHICH_MAX) are Slot Ids and handled above.
Which Ids >= 4000 and <= 4999 are used by EditEngineItemPool,
defined in eeitem.hxx (values 3994 to 4048).

the former hack prevented that some of PoolItems got deleted
when RefCount dropped to 0. when trying to Remove the PoolItem again,
an earlier assert hits.

hack and comment are in source since initial commit,
other referenced comment is nowhere to be found.

Change-Id: I299b2ae4aed088d4ee438397d0bb91ca8d055cef
Reviewed-on: https://gerrit.libreoffice.org/28429
Tested-by: Jochen Nitschke <j.nitschke+logerrit@ok.de>
Reviewed-by: Michael Stahl <mstahl@redhat.com>