Toolbar popups need explicit deactivation
... if based on MenuBarManager, because unlike context
menus, we reuse the same menu instead of creating from
scratch each time, so that it suffers from the same
"double activation" problem as in tdf#97665.

Note that for GtkSalMenu the problem is only for the
top level popup menu, as the Deactivate callback is
properly called when sub menus are closed.

The non-native case also had this problem, but only when
closing the menu by selecting one of its items, and was
fixed already in f0a3d2c9c964f492e5a12ee5f9193748d0b47c65.

Change-Id: Iee037f7baba0f80c32fb3a45f16511f8585c203e