tdf#81345 docx import fix default page break regression
"regression" from 832d620e458f70dbf45ef9d35eebca67d1c6b860

If there are new style elements, then don't treat a
default break in columns as a continuous break.

This fixes both round-tripping, and initial import of
columns and headers on this particular document. Since
MS and LO treat sections so differently, it is a balancing
act of what to change.

Change-Id: I9c79bc1246108da6a5ebbf744acbcfcbb9d33d48
Reviewed-on: https://gerrit.libreoffice.org/28728
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>