tdf#53698: Cache more than 1 item in NumberFormatCodeMapper
In zforlist.cxx the function SvNumberFormatter::GetPreviewStringGuess
switches between the system locale and en_US in order to guess the
formatting of a number.  This results in poor performance, because in
the created SvNumberFormat only the last used locale is cached.

Caching more than 1 entries improves the load time for the document
attached to tdf#53698 from 90s to 60s in case of non en_US locales.

Change-Id: Id0eb4447dea6213015e2d958d8303a1e7892487a
Reviewed-on: https://gerrit.libreoffice.org/29136
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Noel Grandin <noel.grandin@collabora.co.uk>