editeng: invalidate everything only when toggling header-footer edit
The repaining of the full screen can be costly in certain
cases. Here we avoid it altogether if we don't switch
between header/footer editing and document-body editing.

If for whatever reason this results in some regression,
we should then avoid the invalidation for the Online
version only. The change is unconditional because,
in theory, it shouldn't cause regression, and tests
pass.

Change-Id: I05467c98b5fa4857b94f6742004f98dca3799826
Reviewed-on: https://gerrit.libreoffice.org/29272
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>