sw: fix sorting in MarkManager::correctMarksAbsolute()
The recently added tdf90697.rtf caused a libstdc++ assertion about
non-sorted m_vAllMarks.  This is because a SwXParagraph was disposed, a
bookmark on it was moved by correctMarksAbsolute onto the first
paragraph in a table, and there was a SwXTextRange with a UnoMark
pointing to the SwTableNode, so the sort order was not maintained.

So if at least one mark was moved, re-sort the array.

Change-Id: I0efc7827c826e696f29f8480bab1044b31ccfe39