gbuild: try to avoid invoking gdb with LD_LIBRARY_PATH set
Commit c45738dfe0a845834da50dc36652d643c9e1e16d introduced the problem
that gdb will load via LD_LIBRARY_PATH the libpython3.5m.so.1.0 from
INSTDIR instead of the one from the system, which makes it unhappy (at
least on Fedora 24).

Try to hack around that by detecting that the user wants to run gdb, and
passing in a command "set environment LD_LIBRARY_PATH ..." so that "run"
inside gdb uses the custom LD_LIBRARY_PATH but gdb itself does not.

Change-Id: I1d29d2af1a77ffab3e718a876c17c29de444c6d9
Reviewed-on: https://gerrit.libreoffice.org/29349
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Michael Stahl <mstahl@redhat.com>