add complete test for tdf#96453
This is the fourth and final part of the UI test tutorial.

The test combines the previous 3 items in the final test. The original test
from the first tutorial is used as a template, the SvSimpleTable support
is used to assert that the table contains the correct amount of entries,
and the UNO helper methods that the document contains the correct amount
of conditional formats.

Change-Id: Ibe2047e10ff6368c145c88b2bcca9648aa7f8f54