n#780843 revert part of docx testcase
revert testcase for commit d625a749b5b41d93b5a3dae4f95f0299a32eabec

On import, it is impossible to know where automatic
end-of-pages will be, so it cannot be known that all the
content before the page-break will be on the first page.

Removing this testcase because Writer does not have
the capability to dynamically assign headers/footers
to continuous break sections. This test worked on this
particular document, but not on generic documents.

Replacing it with a version of the document with more
text before the continuous break, which produces the
exact opposite footer results.

Change-Id: If660edd47b4eb19798f79babe65c37e1a9848739
Reviewed-on: https://gerrit.libreoffice.org/29408
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>