tdf#102010 Never overwrite MM files via UNO
Fixes the regression introduced by
  commit 9432fe6411963cd9a7fdd32acc5c9a2bf5b9e3e0

This commit removed the crazy bSubjectIsFilename handling,
where I didn't understood the case of bSubjectIsFilename
and a user supplied prefix.

Mail merge to files never overwrites an existing document,
but there is the special case, when a user selects a target
filename in the MM dialog for single file MM.

Should be fixed by a successive commmit, reverting this and
removing an existing file before starting the MM job.

Change-Id: Idda487023e6984de9c1e701fc088a6b7f92e9847