Resolves: tdf#76836 use text format only for text content
Numeric content is not to be formatted with the text format. Entering a
value in such formatted cell still forces it to text content as usual
and then it is displayed using the text format.

The previous behavior displaying also numeric content using the text
format was always wrong and neither Excel nor Gnumeric do it. A numeric
content is displayed left aligned though, which also Excel does.

Change-Id: I3038123b2bfdc4669a8f73840742da92a1df895d