filter: don't loose signing description during pdf export
The free-form string attached to a signature is called description
during ODF/OOXML signing. The certificate chooser dialog has an input
field to provide that.  The PDF export dialog's signature tab reuses
this dialog, but also provides an an own reason input field for the same
purpose.

So in case the generic dialog's description field is filled, don't
simply throw away that string, but set the pdf export's reason field to
the same value.

XDocumentDigitalSignatures.idl is not a published interface and it is
used only internally, so the API change is only nominal.

Change-Id: I6d4cf0b3f586417a76a052dc30c960478a95c984